move consumer-added before pipeline starts playing
As described in https://github.com/centricular/webrtcsink/issues/25#issuecomment-1048465601

Thanks - what I discovered is that things seem to be happy if `create-data-channel` happens before the pipeline starts playing, but `consumer-added` only occurs after the pipeline is playing, and that seems to create a race condition that is easily reproducible with localhost connections. Moving `consumer-added` to just before pipeline starts playing eliminates the issue even with localhost connections.